+++ src/V3EmitC.h
@@ -27,8 +27,9 @@
 
 class V3EmitC final {
 public:
-    static void emitc();
     static void emitcConstPool();
+    static void emitcHeaders();
+    static void emitcImp();
     static void emitcInlines();
     static void emitcModel();
     static void emitcSyms(bool dpiHdrOnly = false);
